GSalary/requirements.txt
aniso8601==9.0.1
Authlib==1.0.1
autopep8==1.6.0
beautifulsoup4==4.11.1
black==22.3.0
boto3==1.23.8
botocore==1.26.8
cachetools==5.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
dnspython==2.2.1
Flask==2.1.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.0
Flask-RESTful==0.3.9
google-auth==2.6.6
google-auth-oauthlib==0.5.1
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
MarkupSafe==2.1.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.29
numpy==1.22.3
nums-from-string==0.1.2
oauthlib==3.2.0
pandas==1.4.2
pathspec==0.9.0
pep8==1.7.1
platformdirs==2.5.2
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
pymongo==4.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.5.2
six==1.16.0
soupsieve==2.3.2.post1
toml==0.10.2
tomli==2.0.1
typing-extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
